requirements.txt
flask
scikit-learn
xgboost
tensorflow
joblib
pandas
matplotlib
seaborn
scipy
Pillow
opencv-python
